Fixed linking issue

+++ inc/phase_info.h
@@ -33,7 +33,7 @@ namespace champsim
 struct phase_info {
   std::string name;
   bool is_warmup;
-  uint64_t length;
+  long long length;
   std::vector<std::size_t> trace_index;
   std::vector<std::string> trace_names;
 };
